requirements.txt
matplotlib==3.6.3
Pillow==11.3.0
skimage==0.0
